refact: moved interfaces
