+++ ui/src/router/routes/open.routes.tsx
@@ -1,6 +1,6 @@
 import { AccessLevel } from "../../enums/AccessLevel";
 import Login from "../../pages/Login";
-import { IAccessRoutes } from "../interfaces";
+import { IAccessRoutes } from "../../interfaces/routes.interfaces";
 
 export default {
     accessLevel: AccessLevel.NONE,
